% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crIntLambda.r
\name{crIntLambda}
\alias{crIntLambda}
\title{Compute the credibility interval for a Poisson model}
\usage{
crIntLambda(x, percentile = c(0.025, 0.975))
}
\arguments{
\item{x}{a vector of observations}

\item{percentile}{a vector of length 2 defining the credibility intervalle lower and upper limits}
}
\value{
a vecteur with the lower and upper limits of the crInt
}
\description{
Compute the credibility interval for a Poisson model
}
\details{
.
}
\examples{
x <- rpois(10, lambda = 3)
crIntLambda(x)
}
\author{
JuG
}
