% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expLambda.r
\name{expLambda}
\alias{expLambda}
\title{Calculate Poisson expectation}
\usage{
expLambda(x)
}
\arguments{
\item{x}{vector of observation of the Poisson generative process}
}
\value{
estimated lambda value
}
\description{
calculate Poisson expectation
}
\examples{
set.seed(123)
x1 <- rpois(10, lambda = 3)
expLambda(x1)
}
\author{
JuG
}
