% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdiLambda.r
\name{hdiLambda}
\alias{hdiLambda}
\title{Compute the HDI for a poisson model}
\usage{
hdiLambda(x, credMass = 0.95, tol = 1e-08, ...)
}
\arguments{
\item{x}{a vector of observations}

\item{credMass}{credibility mass (from 0 to 1)}

\item{tol}{tolerance}
}
\description{
Compute the HDI for a poisson model
}
\examples{
x <- rpois(10, lambda = 3)
hdiLambda(x)
hdiLambda(x = x, credMass = .7)
}
\author{
JuG
}
