% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lim_function.R
\name{lim_function_low}
\alias{lim_function_low}
\title{Calcul de limite inf pour les graphes}
\usage{
lim_function_low(x, offset = 0.3)
}
\arguments{
\item{x}{valeur de la limite basse}

\item{offset}{offset}
}
\description{
Calcul de limite inf pour les graphes
}
