% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lim_function.R
\name{lim_function_up}
\alias{lim_function_up}
\title{Calcul de limite sup pour les graphes}
\usage{
lim_function_up(x, offset = 0.3)
}
\arguments{
\item{x}{valeur de la limite haute}

\item{offset}{offset}
}
\description{
Calcul de limite sup pour les graphes
}
