% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneMeanEstim.R
\name{oneMeanEstim}
\alias{oneMeanEstim}
\title{Estimation d'une moyenne cas conjugé
variable continue
ne peut se faire sans l'estimation de la variance
sauf si supposée connue (=cas rare)}
\usage{
oneMeanEstim(
  X,
  alpha,
  mu_0,
  kappa_0,
  alpha_0,
  beta_0,
  theta_P,
  theta_A,
  seuil = 0
)
}
\arguments{
\item{X}{Variable continue}

\item{alpha}{Pour le calcul de l'IC}

\item{kappa_0}{?}

\item{alpha_0}{Prior sd}

\item{theta_P}{Bornes inf 2IT}

\item{theta_A}{Bornes sup 2IT}

\item{seuil}{mu|sigma2 ~ dnorm(mu_0, sigma2/kappa_0) kappa_0=taille pseudo-échantillon
tau ~ dgamma(alpha_0/2, beta_0/2)
tau ~ dgamma(alpha_n, beta_n)
alpha_n=(alpha_0+n)/2
beta_n=(beta_0+BSSE)/2
BSSE=(n-1)s² + n kappa_0(mu_0 - y.bar)²/(n+kappa_0)
library(ggdist)
mu ~ student(mu_n, v_n, ddl)
mu_n=mu_0 kappa_0/(kapa_0+n) + y.bar n/(kappa_0+n)
v_n=beta_n/alpha_n/(n+kappa_0)
ddl_n=2 alpha_n
Résumé: p50 \link{p2.5; p97.5} mu et sigma=1/sqrt(tau)
Proba de chaque intervalle 2IT
Paramètres de la Student a posteriori de la moyenne
Graphique mu: prior, posterior, bornes IC, bornes 2IT}
}
\value{
Posteriors marginales
}
\description{
Estimation d'une moyenne cas conjugé
variable continue
ne peut se faire sans l'estimation de la variance
sauf si supposée connue (=cas rare)
}
