% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultPlot.R
\name{plot.oneMeanEstim}
\alias{plot.oneMeanEstim}
\title{Dfault plot oneMeanEstim}
\usage{
\method{plot}{oneMeanEstim}(obj, ...)
}
\arguments{
\item{obj}{objet de type oneMeanEstim}

\item{...}{}
}
\description{
Dfault plot oneMeanEstim
}
