% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOneMeanEstim.R
\name{plotOneMeanEstim}
\alias{plotOneMeanEstim}
\title{Graphique oneMeanEstim}
\usage{
plotOneMeanEstim(oneMeanEstimRes, twit = T, area = T, xlim = NULL)
}
\arguments{
\item{oneMeanEstimRes}{object return from oneMeanEstim}

\item{twit}{Boolean to display twit}

\item{area}{Boolean to display twit area}
}
\description{
Graphique oneMeanEstim
}
