% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorPoisson.r
\name{posteriorPoisson}
\alias{posteriorPoisson}
\title{Generate simulated poisson posterior data}
\usage{
posteriorPoisson(x, npoints = 1e+05)
}
\arguments{
\item{x}{a vector of observations}

\item{npoints}{number of points to draw the posterior}
}
\description{
Generate simulated poisson posterior data
}
\examples{
x <- rpois(10, lambda = 3)
hist(posteriorPoisson(x))
}
\author{
JuG
}
