% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultCat.R
\name{print.oneMeanEstim}
\alias{print.oneMeanEstim}
\title{Default cat oneMeanEstim}
\usage{
\method{print}{oneMeanEstim}(x, ...)
}
\arguments{
\item{x}{objet de type oneMeanEstim}

\item{...}{}
}
\description{
Default cat oneMeanEstim
}
