% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneMeanEstim.R
\name{oneMeanEstim}
\alias{oneMeanEstim}
\title{Estimation d'une moyenne cas conjugé
variable continue
ne peut se faire sans l'estimation de la variance
sauf si supposée connue (=cas rare)}
\usage{
oneMeanEstim(
  X,
  mu,
  kappa_0,
  nu,
  sigma_0,
  alpha = 0.05,
  theta_P = NULL,
  theta_A = NULL,
  seuil = 0
)
}
\arguments{
\item{X}{A vector of data values.}

\item{mu}{The prior mean of the population.}

\item{kappa_0}{A value related to the prior precision (inverse variance) of mu}

\item{nu}{A value related to the prior precision of sigman_0}

\item{sigma_0}{The prior sd of standard deviation of the population.}

\item{alpha}{The level of the credibility interval}

\item{theta_P}{Bornes inf 2IT}

\item{theta_A}{Bornes sup 2IT}

\item{seuil}{A threshold value for the probability calculation.
mu|sigma2 ~ dnorm(mu, sigma2/kappa_0) kappa_0=taille pseudo-échantillon
tau ~ dgamma(alpha_0/2, beta_0/2)
tau ~ dgamma(alpha_n, beta_n)
alpha_n=(alpha_0+n)/2
beta_n=(beta_0+BSSE)/2
BSSE=(n-1)s² + n kappa_0(mu - y.bar)²/(n+kappa_0)
library(ggdist)
mu ~ student(mu_n, v_n, ddl)
mu_n=mu kappa_0/(kapa_0+n) + y.bar n/(kappa_0+n)
v_n=beta_n/alpha_n/(n+kappa_0)
ddl_n=2 alpha_n
Résumé: p50 \link{p2.5; p97.5} mu et sigma=1/sqrt(tau)
Proba de chaque intervalle 2IT
Paramètres de la Student a posteriori de la moyenne
Graphique mu: prior, posterior, bornes IC, bornes 2IT}
}
\value{
Posteriors marginales
}
\description{
Estimation d'une moyenne cas conjugé
variable continue
ne peut se faire sans l'estimation de la variance
sauf si supposée connue (=cas rare)
}
