% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOneMeanEstim.R
\name{plotOneMeanEstim}
\alias{plotOneMeanEstim}
\title{Graphique oneMeanEstim}
\usage{
plotOneMeanEstim(oneMeanEstimRes, twoIt = T, area = T, xlim = NULL)
}
\arguments{
\item{oneMeanEstimRes}{object return from oneMeanEstim}

\item{twoIt}{Boolean to display twoIt}

\item{area}{Boolean to display twoIt area}
}
\description{
Graphique oneMeanEstim
}
